Moved setting panels into separate folder
