+++ frontend/src/components/settingPanels/CheckboxSetting.vue
