+++ frontend/src/components/settingPanels/ColorSetting.vue
