+++ frontend/src/components/settingPanels/NumberSetting.vue
